+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * added tuple for session state cancellation
 * removed the requirement to send a page view before an event can be tracked
+* fixed grouping visitors by hour
 
 ## 3.4.0